# openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "photo-contest",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"variables": {}
		}
	],
	"paths": {
...
		"/votes/votesByPhoto": {
			"get": {
				"tags": ["votes"],
				"summary": "getAllVotes",
				"operationId": "getAllVotes",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/votes": {
			"post": {
				"tags": ["votes"],
				"summary": "addVote",
				"operationId": "addVote",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/addVoteRequest"
									},
									{
										"example": {
											"user_id": 2,
											"photo_id": 1,
											"caption_id": 1
										}
									}
								]
							},
							"example": {
								"id": 4,
								"user_id": 2,
								"photo_id": 1,
								"caption_id": 1
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"addVoteRequest": {
				"title": "addVoteRequest",
				"required": ["id", "user_id", "photo_id", "caption_id"],
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"user_id": {
						"type": "integer",
						"format": "int32"
					},
					"photo_id": {
						"type": "integer",
						"format": "int32"
					},
					"caption_id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"user_id": 8,
					"photo_id": 1,
					"caption_id": 1
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}